chore(remove): event-hubs pump project
