--- src/Arcus.Messaging.Pumps.EventHubs/Arcus.Messaging.Pumps.EventHubs.csproj
@@ -1,41 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net8.0;net6.0;netstandard2.1</TargetFrameworks>
-    <LangVersion>latest</LangVersion>
-    <Authors>Arcus</Authors>
-    <Company>Arcus</Company>
-    <Product>Arcus.Messaging</Product>
-    <Description>Provides capability to run an Azure Event Hubs message pump</Description>
-    <Copyright>Copyright (c) Arcus</Copyright>
-    <PackageProjectUrl>https://messaging.arcus-azure.net/</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/arcus-azure/arcus.messaging</RepositoryUrl>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <PackageIcon>icon.png</PackageIcon>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <RepositoryType>Git</RepositoryType>
-    <PackageTags>Azure;Messaging;EventHubs</PackageTags>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\..\README.md" Pack="true" PackagePath="\" />
-    <None Include="..\..\LICENSE" Pack="true" PackagePath="\" />
-    <None Include="..\..\docs\static\img\icon.png" Pack="true" PackagePath="\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Arcus.Security.Core" Version="[2.0.0,3.0.0)" />
-    <PackageReference Include="Azure.Messaging.EventHubs" Version="5.11.3" />
-    <PackageReference Include="Azure.Messaging.EventHubs.Processor" Version="5.11.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Arcus.Messaging.Abstractions.EventHubs\Arcus.Messaging.Abstractions.EventHubs.csproj" />
-    <ProjectReference Include="..\Arcus.Messaging.EventHubs.Core\Arcus.Messaging.EventHubs.Core.csproj" />
-    <ProjectReference Include="..\Arcus.Messaging.Pumps.Abstractions\Arcus.Messaging.Pumps.Abstractions.csproj" />
-  </ItemGroup>
-
-</Project>
