+++ src/Arcus.Messaging.sln
@@ -29,8 +29,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Messaging.Abstraction
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Messaging.AzureFunctions.ServiceBus", "Arcus.Messaging.AzureFunctions.ServiceBus\Arcus.Messaging.AzureFunctions.ServiceBus.csproj", "{7386C41A-6F27-42B7-BBC1-B6CB41200A68}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Messaging.Pumps.EventHubs", "Arcus.Messaging.Pumps.EventHubs\Arcus.Messaging.Pumps.EventHubs.csproj", "{F5A14425-FEF5-425D-875C-2601C988E6FA}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Messaging.Abstractions.EventHubs", "Arcus.Messaging.Abstractions.EventHubs\Arcus.Messaging.Abstractions.EventHubs.csproj", "{040DE2D5-0400-4B4A-A333-60C45CBB0204}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Messaging.EventHubs.Core", "Arcus.Messaging.EventHubs.Core\Arcus.Messaging.EventHubs.Core.csproj", "{05EA4B3A-C619-44BB-88C4-F067FCCF83AA}"
@@ -99,10 +97,6 @@ Global
 		{7386C41A-6F27-42B7-BBC1-B6CB41200A68}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7386C41A-6F27-42B7-BBC1-B6CB41200A68}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7386C41A-6F27-42B7-BBC1-B6CB41200A68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F5A14425-FEF5-425D-875C-2601C988E6FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F5A14425-FEF5-425D-875C-2601C988E6FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F5A14425-FEF5-425D-875C-2601C988E6FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F5A14425-FEF5-425D-875C-2601C988E6FA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{040DE2D5-0400-4B4A-A333-60C45CBB0204}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{040DE2D5-0400-4B4A-A333-60C45CBB0204}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{040DE2D5-0400-4B4A-A333-60C45CBB0204}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -137,7 +131,6 @@ Global
 		{55DE6D12-4C54-4570-BDFA-00B0FFDE5AB6} = {A1369CCD-42D1-43F6-98BC-D8EDA62C2B13}
 		{864C12DF-DE3D-421F-8687-EC3918FFB8BE} = {2CD090E7-7306-49A0-9680-6ED78CFECAE1}
 		{7386C41A-6F27-42B7-BBC1-B6CB41200A68} = {2CD090E7-7306-49A0-9680-6ED78CFECAE1}
-		{F5A14425-FEF5-425D-875C-2601C988E6FA} = {31250C96-0F22-482F-B3D7-127898A28CF3}
 		{040DE2D5-0400-4B4A-A333-60C45CBB0204} = {31250C96-0F22-482F-B3D7-127898A28CF3}
 		{05EA4B3A-C619-44BB-88C4-F067FCCF83AA} = {31250C96-0F22-482F-B3D7-127898A28CF3}
 		{F1B24FD4-099E-489D-B45D-A1A992CA5C3A} = {A1369CCD-42D1-43F6-98BC-D8EDA62C2B13}
